backend/requirements.txt
annotated-types==0.6.0
anyio==4.4.0
asyncio==3.4.3
attrs==20.3.0
autopep8==2.1.0
backports.shutil-get-terminal-size==1.0.0
backports.shutil_which==3.5.2
bcrypt==4.1.3
beautifulsoup4==4.12.3
black==24.3.0
blinker==1.7.0
braceexpand==0.1.7
Brotli==1.1.0
cachelib==0.12.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
check50==3.3.11
clang-format==18.1.2
cli50==7.5.1
click==8.1.7
colorama==0.4.6
compare50==1.2.6
contourpy==1.2.0
cryptography==42.0.5
cs50==9.3.4
cssselect2==0.7.0
cycler==0.12.1
distro==1.9.0
djhtml==3.0.6
EditorConfig==0.12.4
Flask==3.0.2
Flask-Cors==4.0.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
fonttools==4.50.0
greenlet==3.0.3
h11==0.14.0
help50==3.0.5
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
icdiff==2.0.7
idna==3.6
inflect==7.0.0
iniconfig==2.0.0
intervaltree==3.1.0
itsdangerous==2.1.2
jellyfish==0.11.2
Jinja2==3.1.3
jsbeautifier==1.15.1
kiwisolver==1.4.5
lib50==3.0.11
MarkupSafe==2.1.5
matplotlib==3.8.3
msgspec==0.18.6
mypy-extensions==1.0.0
natsort==8.4.0
numpy==1.26.4
openai==1.35.7
packaging==24.0
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
psutil==5.9.8
ptyprocess==0.7.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.16.3
pydyf==0.9.0
Pygments==2.17.2
pyparsing==3.1.2
pypdf==4.1.0
pyphen==0.14.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
render50==9.2.7
requests==2.31.0
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.4.4
style50==2.7.6
submit50==3.1.4
termcolor==1.1.0
tinycss2==1.2.1
tqdm==4.66.2
typing_extensions==4.10.0
tzlocal==5.2
urllib3==2.2.1
weasyprint==61.2
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
wheel==0.43.0
zopfli==0.2.3
